add Async.when
Async.when can be though of as something like Future.map_result().
This is useful for when you need to compute a Future based of the ready
result of a previous Future, and need to move a specific value into it.
The documentation includes an example that cannot be solved with
`and_then` and `or_else` together.